Move ssrExternals into config directory
