+++ packages/react-router-dev/config/config.ts
@@ -22,7 +22,6 @@ import {
   configRoutesToRouteManifest,
 } from "./routes";
 import { detectPackageManager } from "../cli/detectPackageManager";
-import { ssrExternals } from "../vite/ssr-externals";
 
 const excludedConfigPresetKeys = ["presets"] as const satisfies ReadonlyArray<
   keyof ReactRouterConfig
@@ -700,6 +699,34 @@ export async function resolveEntryFiles({
     : path.resolve(defaultsDirectory, entryServerFile);
 
   return { entryClientFilePath, entryServerFilePath };
+}
+
+export const ssrExternals = isInReactRouterMonorepo()
+  ? [
+      // This is only needed within this repo because these packages
+      // are linked to a directory outside of node_modules so Vite
+      // treats them as internal code by default.
+      "react-router",
+      "react-router-dom",
+      "@react-router/architect",
+      "@react-router/cloudflare",
+      "@react-router/dev",
+      "@react-router/express",
+      "@react-router/node",
+      "@react-router/serve",
+    ]
+  : undefined;
+
+function isInReactRouterMonorepo() {
+  // We use '@react-router/node' for this check since it's a
+  // dependency of this package and guaranteed to be in node_modules
+  let serverRuntimePath = path.dirname(
+    require.resolve("@react-router/node/package.json")
+  );
+  let serverRuntimeParentDir = path.basename(
+    path.resolve(serverRuntimePath, "..")
+  );
+  return serverRuntimeParentDir === "packages";
 }
 
 const entryExts = [".js", ".jsx", ".ts", ".tsx"];
